refactor: cp keepers and queries (#79)
* proto

* refactor query

* refactor

* clean

* refactor

* fix

* refactor

* revert

* revert

* revert

* refactor

* revert

* revert

* validate

* integ

* fix

* clean

* move ctx

* fix

---------

+++ pkg/math/math.go
@@ -20,7 +20,7 @@ func Min[V int | int64 | uint64 | int32 | uint32](vals ...V) V {
 }
 
 // Float64StringToBigInt converts a float64 string to a big.Int.
-func Float64StringToBigInt(s string, decimals int) (*big.Int, error) {
+func Float64StringToBigInt(s string, decimals int64) (*big.Int, error) {
 	bigFloat := new(big.Float)
 	_, _, err := bigFloat.Parse(s, 10)
 	if err != nil {
@@ -31,7 +31,7 @@ func Float64StringToBigInt(s string, decimals int) (*big.Int, error) {
 }
 
 // Float64ToBigInt converts a float64 to a big.Int.
-func Float64ToBigInt(val float64, decimals int) *big.Int {
+func Float64ToBigInt(val float64, decimals int64) *big.Int {
 	bigVal := new(big.Float)
 	bigVal.SetFloat64(val)
 
@@ -39,7 +39,7 @@ func Float64ToBigInt(val float64, decimals int) *big.Int {
 }
 
 // BigFloatToBigInt converts a big.Float to a big.Int.
-func BigFloatToBigInt(f *big.Float, decimals int) *big.Int {
+func BigFloatToBigInt(f *big.Float, decimals int64) *big.Int {
 	bigFloat := new(big.Float)
 	factor := big.NewInt(1).Exp(big.NewInt(10), big.NewInt(int64(decimals)), nil)
 	bigFloat.SetInt(factor)
